pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "whatsapp-chatbot-api"
version = "1.0.0"
description = "Ultra-lightweight WhatsApp Chat Query API using FastAPI and Vercel"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.104.0,<0.105.0",
    "uvicorn[standard]>=0.24.0,<0.25.0",
    "python-multipart>=0.0.6,<0.1.0",
    "pydantic>=2.5.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0"
]

[tool.uv]
package-manager = "uv"
